use <cstdint> types

+++ mortonkeys.cpp
@@ -1,15 +1,16 @@
 
 #include <iostream>
+#include <cstdint>
 #include <vector>
 
 #define MAXMORTONKEY 1317624576693539401 //21 bits MAX morton key
 
-std::vector<unsigned long long> mortonkeyX;
-std::vector<unsigned long long> mortonkeyY;
-std::vector<unsigned long long> mortonkeyZ;
+std::vector<uint64_t> mortonkeyX;
+std::vector<uint64_t> mortonkeyY;
+std::vector<uint64_t> mortonkeyZ;
 
 /*Compute a 256 array of morton keys at compile time*/
-template <unsigned int i, unsigned int offset>
+template <uint32_t i, uint32_t offset>
 struct morton
 {
 	enum
@@ -18,28 +19,28 @@ struct morton
 		//mortonkey(x+1) = (mortonkey(x) - MAXMORTONKEY) & MORTONKEY
 		value = (morton<i - 1, offset>::value - MAXMORTONKEY) & MAXMORTONKEY
 	};
-	static void add_values(std::vector<unsigned long long>& v)
+	static void add_values(std::vector<uint64_t>& v)
 	{
 		morton<i - 1, offset>::add_values(v);
 		v.push_back(value<<offset);
 	}
 };
 
-template <unsigned int offset>
+template <uint32_t offset>
 struct morton<0, offset>
 {
 	enum
 	{
 		value = 0
 	};
-	static void add_values(std::vector<unsigned long long>& v)
+	static void add_values(std::vector<uint64_t>& v)
 	{
 		v.push_back(value);
 	}
 };
 
-inline unsigned long long encodeMortonKey(unsigned int x, unsigned int y, unsigned int z){
-	unsigned long long result = 0;
+inline uint64_t encodeMortonKey(uint32_t x, uint32_t y, uint32_t z){
+	uint64_t result = 0;
 	result = mortonkeyZ[(z >> 16) & 0xFF] |
 		mortonkeyY[(y >> 16) & 0xFF] |
 		mortonkeyX[(x >> 16) & 0xFF];
@@ -63,7 +64,7 @@ int main()
 	morton<256, 1>::add_values(mortonkeyY);
 	morton<256, 2>::add_values(mortonkeyZ);
 
-	unsigned long long key;
+	uint64_t key;
 
 	for (int x = 0; x <= TESTSIZE; ++x)
 	  for (int y = 0; y <= TESTSIZE; ++y)
